Better article definitions
