+++ src/assets/markdown/Welcome.md
@@ -1,0 +1,1 @@
+Welcome to Aeirth!
